Renaming to storywright (#4)
