+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from './StoryRunner/Steps';
-export * from './StoryRunner/StoryRunner';
-export * from './StoryRunner/Keys';
+export * from './StoryWright/Steps';
+export * from './StoryWright/StoryRunner';
+export * from './StoryWright/Keys';
